@@ -1,71 +1,45 @@
-/*
- *  File:   DSharpUtil.cs
- *  Author: Angelo Breuer
- *
- *  The MIT License (MIT)
- *
- *  Copyright (c) Angelo Breuer 2022
- *
- *  Permission is hereby granted, free of charge, to any person obtaining a copy
- *  of this software and associated documentation files (the "Software"), to deal
- *  in the Software without restriction, including without limitation the rights
- *  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- *  copies of the Software, and to permit persons to whom the Software is
- *  furnished to do so, subject to the following conditions:
- *
- *  The above copyright notice and this permission notice shall be included in
- *  all copies or substantial portions of the Software.
- *
- *  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- *  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- *  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- *  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- *  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- *  THE SOFTWARE.
- */
+namespace Lavalink4NET.DSharpPlus;
 
-namespace Lavalink4NET.DSharpPlus
+using System.Reflection;
+using global::DSharpPlus;
+using global::DSharpPlus.Entities;
+using global::DSharpPlus.Net.WebSocket;
+
+/// <summary>
+///     An utility for getting internal / private fields from DSharpPlus WebSocket Gateway Payloads.
+/// </summary>
+public static class DSharpUtil
 {
-    using System.Reflection;
-    using global::DSharpPlus;
-    using global::DSharpPlus.Entities;
-    using global::DSharpPlus.EventArgs;
-    using global::DSharpPlus.Net.WebSocket;
+    /// <summary>
+    ///     The internal "SessionId" property info in <see cref="DiscordVoiceState"/>.
+    /// </summary>
+    // https://github.com/DSharpPlus/DSharpPlus/blob/master/DSharpPlus/Entities/Voice/DiscordVoiceState.cs#L95
+    private static readonly PropertyInfo _sessionIdProperty = typeof(DiscordVoiceState)
+#pragma warning disable S3011 // Reflection should not be used to increase accessibility of classes, methods, or fields
+        .GetProperty("SessionId", BindingFlags.NonPublic | BindingFlags.Instance)!;
+#pragma warning restore S3011 // Reflection should not be used to increase accessibility of classes, methods, or fields
+
 
     /// <summary>
-    ///     An utility for getting internal / private fields from DSharpPlus WebSocket Gateway Payloads.
+    ///     The internal "_webSocketClient" field info in <see cref="DiscordClient"/>.
     /// </summary>
-    public static class DSharpUtil
-    {
-        /// <summary>
-        ///     The internal "SessionId" property info in <see cref="DiscordVoiceState"/>.
-        /// </summary>
-        // https://github.com/DSharpPlus/DSharpPlus/blob/master/DSharpPlus/Entities/Voice/DiscordVoiceState.cs#L95
-        private static readonly PropertyInfo _sessionIdProperty = typeof(DiscordVoiceState)
-            .GetProperty("SessionId", BindingFlags.NonPublic | BindingFlags.Instance);
+    // https://github.com/DSharpPlus/DSharpPlus/blob/master/DSharpPlus/Clients/DiscordClient.WebSocket.cs#L54
+    private static readonly FieldInfo _webSocketClientField = typeof(DiscordClient)
+#pragma warning disable S3011 // Reflection should not be used to increase accessibility of classes, methods, or fields
+        .GetField("_webSocketClient", BindingFlags.NonPublic | BindingFlags.Instance)!;
+#pragma warning restore S3011 // Reflection should not be used to increase accessibility of classes, methods, or fields
 
-     
-        /// <summary>
-        ///     The internal "_webSocketClient" field info in <see cref="DiscordClient"/>.
-        /// </summary>
-        // https://github.com/DSharpPlus/DSharpPlus/blob/master/DSharpPlus/Clients/DiscordClient.WebSocket.cs#L54
-        private static readonly FieldInfo _webSocketClientField = typeof(DiscordClient)
-            .GetField("_webSocketClient", BindingFlags.NonPublic | BindingFlags.Instance);
+    /// <summary>
+    ///     Gets the internal "SessionId" property value of the specified <paramref name="voiceState"/>.
+    /// </summary>
+    /// <param name="voiceState">the instance</param>
+    /// <returns>the "SessionId" value</returns>
+    public static string GetSessionId(this DiscordVoiceState voiceState)
+        => (string)_sessionIdProperty.GetValue(voiceState)!;
 
-        /// <summary>
-        ///     Gets the internal "SessionId" property value of the specified <paramref name="voiceState"/>.
-        /// </summary>
-        /// <param name="voiceState">the instance</param>
-        /// <returns>the "SessionId" value</returns>
-        public static string GetSessionId(this DiscordVoiceState voiceState)
-            => (string)_sessionIdProperty.GetValue(voiceState);
-
-
-        /// <summary>
-        ///     Gets the internal "_webSocketClient" field value of the specified <paramref name="client"/>.
-        /// </summary>
-        public static IWebSocketClient GetWebSocketClient(this DiscordClient client)
-            => (IWebSocketClient)_webSocketClientField.GetValue(client);
-    }
+    /// <summary>
+    ///     Gets the internal "_webSocketClient" field value of the specified <paramref name="client"/>.
+    /// </summary>
+    public static IWebSocketClient GetWebSocketClient(this DiscordClient client)
+        => (IWebSocketClient)_webSocketClientField.GetValue(client)!;
 }
