@@ -1,16 +1,20 @@
 ﻿namespace Lavalink4NET.Player;
 
+using System;
 using Lavalink4NET.Integrations.ExtraFilters;
+using Lavalink4NET.Players;
 
 public static class PlayerFilterMapExtensions
 {
-    public static void Echo(this PlayerFilterMap filterMap, EchoFilterOptions filterOptions)
+    public static void Echo(this IPlayerFilters filters, EchoFilterOptions? filterOptions)
     {
-        filterMap[EchoFilterOptions.Name] = filterOptions;
+        ArgumentNullException.ThrowIfNull(filters);
+        filters.SetFilter(filterOptions);
     }
 
-    public static EchoFilterOptions? Echo(this PlayerFilterMap filterMap)
+    public static EchoFilterOptions? Echo(this IPlayerFilters filters)
     {
-        return filterMap[EchoFilterOptions.Name] as EchoFilterOptions;
+        ArgumentNullException.ThrowIfNull(filters);
+        return filters.GetFilter<EchoFilterOptions>();
     }
 }
