@@ -1,51 +1,30 @@
-/*
- *  File:   ChannelMixFilterOptions.cs
- *  Author: Angelo Breuer
- *
- *  The MIT License (MIT)
- *
- *  Copyright (c) Angelo Breuer 2022
- *
- *  Permission is hereby granted, free of charge, to any person obtaining a copy
- *  of this software and associated documentation files (the "Software"), to deal
- *  in the Software without restriction, including without limitation the rights
- *  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- *  copies of the Software, and to permit persons to whom the Software is
- *  furnished to do so, subject to the following conditions:
- *
- *  The above copyright notice and this permission notice shall be included in
- *  all copies or substantial portions of the Software.
- *
- *  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- *  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- *  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- *  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- *  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- *  THE SOFTWARE.
- */
-
 namespace Lavalink4NET.Filters;
 
-using System.Text.Json.Serialization;
+using Lavalink4NET.Protocol.Models.Filters;
 
-
-public sealed class ChannelMixFilterOptions : IFilterOptions
+public sealed record class ChannelMixFilterOptions(
+    float? LeftToLeft = null,
+    float? LeftToRight = null,
+    float? RightToLeft = null,
+    float? RightToRight = null) : IFilterOptions
 {
-    public const string Name = "channelMix";
+    public bool IsDefault => this is
+    {
+        LeftToLeft: 1F or null,
+        LeftToRight: 0F or null,
+        RightToLeft: 0F or null,
+        RightToRight: 1F or null,
+    };
 
-    /// <inheritdoc/>
-    string IFilterOptions.Name => Name;
-
-    [JsonPropertyName("leftToLeft")]
-    public float LeftToLeft { get; init; } = 1F;
-
-    [JsonPropertyName("leftToRight")]
-    public float LeftToRight { get; init; } = 0F;
-
-    [JsonPropertyName("rightToLeft")]
-    public float RightToLeft { get; init; } = 0F;
-
-    [JsonPropertyName("rightToRight")]
-    public float RightToRight { get; init; } = 1F;
+    public void Apply(ref PlayerFilterMapModel filterMap)
+    {
+        filterMap = filterMap with
+        {
+            ChannelMix = new ChannelMixFilterModel(
+                LeftToLeft: LeftToLeft,
+                LeftToRight: LeftToRight,
+                RightToLeft: RightToLeft,
+                RightToRight: RightToRight),
+        };
+    }
 }
