@@ -1,48 +1,24 @@
-/*
- *  File:   TimescaleFilterOptions.cs
- *  Author: Angelo Breuer
- *
- *  The MIT License (MIT)
- *
- *  Copyright (c) Angelo Breuer 2022
- *
- *  Permission is hereby granted, free of charge, to any person obtaining a copy
- *  of this software and associated documentation files (the "Software"), to deal
- *  in the Software without restriction, including without limitation the rights
- *  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- *  copies of the Software, and to permit persons to whom the Software is
- *  furnished to do so, subject to the following conditions:
- *
- *  The above copyright notice and this permission notice shall be included in
- *  all copies or substantial portions of the Software.
- *
- *  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- *  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- *  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- *  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- *  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- *  THE SOFTWARE.
- */
-
 namespace Lavalink4NET.Filters;
 
-using System.Text.Json.Serialization;
+using Lavalink4NET.Protocol.Models.Filters;
 
-
-public sealed class TimescaleFilterOptions : IFilterOptions
+public sealed record class TimescaleFilterOptions(
+    float? Speed = null,
+    float? Pitch = null,
+    float? Rate = null) : IFilterOptions
 {
-    public const string Name = "timescale";
+    public bool IsDefault => this is
+    {
+        Speed: null or 1.0F,
+        Pitch: null or 1.0F,
+        Rate: null or 1.0F,
+    };
 
-    /// <inheritdoc/>
-    string IFilterOptions.Name => Name;
-
-    [JsonPropertyName("speed")]
-    public float Speed { get; set; } = 1.0F;
-
-    [JsonPropertyName("pitch")]
-    public float Pitch { get; set; } = 1.0F;
-
-    [JsonPropertyName("rate")]
-    public float Rate { get; set; } = 1.0F;
+    public void Apply(ref PlayerFilterMapModel filterMap)
+    {
+        filterMap = filterMap with
+        {
+            Timescale = new TimescaleFilterModel(Speed, Pitch, Rate),
+        };
+    }
 }
